+++ tests/integration/cases/record_container_fields/CommonLisp@ansi.lisp
@@ -1,0 +1,4 @@
+(defparameter *my_data* (list
+    (list (cons "id" 1) (cons "empty_map" nil) (cons "int_map" (list (cons 1 "a"))) (cons "full_set" (list "x" "y")) (cons "empty_set" nil))
+    (list (cons "id" 2) (cons "empty_map" nil) (cons "int_map" (list (cons 1 "b"))) (cons "full_set" (list "x")) (cons "empty_set" nil))
+))