+++ tests/integration/cases/record_container_fields/Groovy@v4.groovy
@@ -1,0 +1,4 @@
+def my_data = [
+    ["id": 1, "empty_map": [:], "int_map": [1: "a"], "full_set": ["x", "y"] as Set<Object>, "empty_set": [] as Set<Object>],
+    ["id": 2, "empty_map": [:], "int_map": [1: "b"], "full_set": ["x"] as Set<Object>, "empty_set": [] as Set<Object>],
+]