+++ tests/integration/cases/record_container_fields/Perl_combined@v5_36.pl
@@ -1,0 +1,8 @@
+my $my_data = [
+    {"id" => 1, "empty_map" => {}, "int_map" => {1 => "a"}, "full_set" => ["x", "y"], "empty_set" => []},
+    {"id" => 2, "empty_map" => {}, "int_map" => {1 => "b"}, "full_set" => ["x"], "empty_set" => []},
+];
+$my_data = [
+    {"id" => 1, "empty_map" => {}, "int_map" => {1 => "a"}, "full_set" => ["x", "y"], "empty_set" => []},
+    {"id" => 2, "empty_map" => {}, "int_map" => {1 => "b"}, "full_set" => ["x"], "empty_set" => []},
+];