+++ tests/integration/cases/record_container_fields/PowerShell@v7.ps1
@@ -1,0 +1,4 @@
+$my_data = @(
+    @{"id" = 1; "empty_map" = @{}; "int_map" = @{1 = "a"}; "full_set" = @("x"; "y"); "empty_set" = @()};
+    @{"id" = 2; "empty_map" = @{}; "int_map" = @{1 = "b"}; "full_set" = @("x"); "empty_set" = @()}
+)