+++ tests/integration/cases/record_container_fields/Scheme@r7rs.scm
@@ -1,0 +1,4 @@
+(define my_data (list
+    (list "id" 1 "empty_map" (list) "int_map" (list 1 "a") "full_set" (list "x" "y") "empty_set" (list))
+    (list "id" 2 "empty_map" (list) "int_map" (list 1 "b") "full_set" (list "x") "empty_set" (list))
+))