+++ tests/integration/cases/record_container_fields/Swift_combined@v5_9.swift
@@ -1,0 +1,8 @@
+var my_data: Any = [
+    ["id": 1, "empty_map": [String: Any](), "int_map": [1: "a"], "full_set": Set<AnyHashable>(["x", "y"]), "empty_set": Set<AnyHashable>()],
+    ["id": 2, "empty_map": [String: Any](), "int_map": [1: "b"], "full_set": Set<AnyHashable>(["x"]), "empty_set": Set<AnyHashable>()],
+]
+my_data = [
+    ["id": 1, "empty_map": [String: Any](), "int_map": [1: "a"], "full_set": Set<AnyHashable>(["x", "y"]), "empty_set": Set<AnyHashable>()],
+    ["id": 2, "empty_map": [String: Any](), "int_map": [1: "b"], "full_set": Set<AnyHashable>(["x"]), "empty_set": Set<AnyHashable>()],
+]