+++ tests/integration/cases/record_container_fields/Tcl_combined@v8_6.tcl
@@ -1,0 +1,8 @@
+set my_data [list \
+    [dict create "id" 1 "empty_map" [dict create ] "int_map" [dict create 1 "a"] "full_set" [dict create "x" 1 "y" 1] "empty_set" [dict create ]] \
+    [dict create "id" 2 "empty_map" [dict create ] "int_map" [dict create 1 "b"] "full_set" [dict create "x" 1] "empty_set" [dict create ]] \
+]
+set my_data [list \
+    [dict create "id" 1 "empty_map" [dict create ] "int_map" [dict create 1 "a"] "full_set" [dict create "x" 1 "y" 1] "empty_set" [dict create ]] \
+    [dict create "id" 2 "empty_map" [dict create ] "int_map" [dict create 1 "b"] "full_set" [dict create "x" 1] "empty_set" [dict create ]] \
+]